refactor: move template generation to deployment
