+++ cli/src/main/java/io/quarkiverse/backstage/cli/template/UninstallCommand.java
@@ -48,7 +48,7 @@ public class UninstallCommand extends GenerationBaseCommand {
             try {
                 refreshed = getBackstageClient().getEntity(entity.getKind().toLowerCase(),
                         entity.getMetadata().getNamespace().orElse("default"), entity.getMetadata().getName());
-                String locationTarget = refreshed.getMetadata().getAnnotations().get("backstage.io/managed-by-location")
+                String locationTarget = refreshed.getMetadata().getAnnotations().get("backstage.io/managed-by-origin-location")
                         .replaceAll("url:", "")
                         .replaceAll("file:", "");
